fix load audio for tfio.audio

+++ vistec_ser/data/features/preprocessing.py
@@ -18,7 +18,7 @@ def load_waveform(
     # elif '.flac' in audio:
     #     waveform = tf.divide(tfio.audio.decode_flac(audio_binary, dtype=tf.int16), normalize_factor)
     audio_tensor = tfio.audio.AudioIOTensor(filename=audio, dtype=tf.int64)
-    audio_rate = audio_tensor.rate
+    audio_rate = tf.cast(audio_tensor.rate, tf.int64)
     audio_tensor = tf.cast(audio_tensor.to_tensor(), tf.float32)
     waveform = tf.divide(audio_tensor, normalize_factor)
     waveform = tf.squeeze(waveform, axis=-1)
